Move some pydantic models to lib/galaxy/schema/schema.py
This should simplify the imports and will allow to reuse the models or decompose them more easily
